fix(agent-cli): strip app_notification wrapper inbound and pin its proof tag
Round-3 AC-7 taught the outbound fence walk to tag a notification's
title/body app_notification instead of job_posting, but the inbound
unfence mirror still stripped only job_posting for those two field
names -- a caller echoing a fenced notification title back into a
write persisted the literal wrapper into the user's own store. Both
strips now run (a no-op on the other tag's wrapper, same as the
existing text field's origin split).

The proof-vs-fencing regression test built its ListMatch fixture
without the createdAt/read anchor a real notifications_list row
always carries, so it never actually exercised the app_notification
branch; production always saw a non-notification-shaped fixture and
the test's hardcoded job_posting expectation happened to match. The
fixture now carries that anchor for a notifications_list-backed row,
and the expected tag is derived per-row so the fenced branch asserts
the tag production actually emits.

Added agent_call::reshape::EMITTED_FENCE_TAGS, the surface's own
hand-audited list of every tag it can pass to prompt_fence::fenced,
and a test pinning that mcp instructions documents each one -- the
crate-wide EXPECTED_FENCE_TAGS registry covers tags this surface never
emits and can't stand in as its coverage source.

Moved proof.rs's inline test module to proof/tests.rs (same split
validate.rs already uses) -- the fixture/assertion additions pushed it
past R8's hard LOC cap.

Refs #1157, #1162
